Fixed terms and interfaces definitions and links
This update ensures that all terms in IDL blocks link to the prose that defines
them in the spec, and updates several references to terms.

The "establish a connection with the remote playback device" algorithm was not
referenced from anywhere. I added to the connect() algorithm prior to starting
remote playback.