Adds customized paragraph to SoTD that mentions the Presentation API
This update intents to fulfill the following pubrules SoTD requirement:
"It MUST include at least one customized paragraph. This section SHOULD include
the title page date (i.e., the one next to the maturity level at the top of the
document). These paragraphs SHOULD explain the publication context, including
rationale and relationships to other work."
https://www.w3.org/pubrules/doc/rules/?profile=FPWD#customParagraph

While the Remote Playback API does not depend on the Presentation API in
practice, it seems important to present it in the context of the work on the
Presentation API.