Fix "allowed to show a popup" concept
The "allowed to show a popup" concept no longer exist in HTML and thus the link
to "allowed to show a popup" is now a broken link.

Tracking user activation is to be done through the concepts of sticky/transient
activation:
https://html.spec.whatwg.org/multipage/interaction.html#user-activation-data-model

This update replaces "isn't allowed to show a popup" with "does not have
transient activation".