Add crEnd date to please ReSpec
ReSpec requires a `crEnd` date when the spec's status is CR or CRD. It makes
sense for CR (the date appears in the Status of this Document section). It does
not really make sense for CRD (the date does not appear anywhere). Anyway,
adding the parameter to please the ReSpec and make the auto-publication process
succeed.